Deadpool support (missing tests)
