+++ Cargo.toml
@@ -33,6 +33,7 @@ chrono = { version = "0.4.25", default-features = false, features = [
   "std",
 ] }
 crossbeam-channel = "0.5"
+deadpool = "0.12"
 env_logger = "0.11"
 fake = "2.6"
 log = "0.4"
