+++ butane/Cargo.toml
@@ -15,6 +15,7 @@ build = "build.rs"
 [features]
 async = ["butane_core/async", "butane_codegen/async"]
 async-adapter = ["butane_core/async-adapter"]
+deadpool = ["dep:deadpool", "async"]
 default = ["datetime", "json", "uuid"]
 fake = ["butane_core/fake"]
 json = ["butane_codegen/json", "butane_core/json"]
@@ -24,13 +25,15 @@ pg = ["async", "butane_core/pg"]
 datetime = ["butane_codegen/datetime", "butane_core/datetime"]
 debug = ["butane_core/debug"]
 log = ["butane_core/log"]
-r2d2 = ["butane_core/r2d2"]
+r2d2 = ["dep:r2d2"]
 tls = ["butane_core/tls"]
 uuid = ["butane_codegen/uuid", "butane_core/uuid"]
 
 [dependencies]
 butane_codegen = { workspace = true }
 butane_core = { workspace = true }
+r2d2 = { optional = true, workspace = true }
+deadpool = { optional = true, workspace = true }
 
 [dev-dependencies]
 butane_test_helper = { workspace = true }
@@ -51,7 +54,6 @@ tokio = { workspace = true, features = ["macros"] }
 tokio-postgres = { features = ["with-geo-types-0_7"], workspace = true }
 tokio-test = { workspace = true }
 rand = { workspace = true }
-r2d2_for_test = { package = "r2d2", version = "0.8" }
 rusqlite = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
